+++ com.unity.mobile.notifications/Runtime/Android/AndroidNotification.cs
@@ -13,11 +13,10 @@ namespace Unity.Notifications.Android
         /// </summary>
         None = 0,
 
-        //// todo currently disabled, bigpicture style requires additional logic that will be implemented in a future release
-        ///// <summary>
-        ///// generate a large-format notification.
-        ///// </summary>
-        //bigpicture = 1,
+        /// <summary>
+        /// Generate a large-format notification centered around an image.
+        /// </summary>
+        BigPictureStyle = 1,
 
         /// <summary>
         /// Generate a large-format notification that includes a lot of text.
@@ -49,6 +48,46 @@ namespace Unity.Notifications.Android
     }
 
     /// <summary>
+    /// Data for setting up the big picture style notification.
+    /// Properties that are not available in devices API level are ignored. See Android documentation for availibility.
+    /// </summary>
+    /// <see cref="https://developer.android.com/reference/android/app/Notification.BigPictureStyle"/>
+    public struct BigPictureStyle
+    {
+        /// <summary>
+        /// The override for large icon (requirements are the same).
+        /// </summary>
+        /// <see cref="AndroidNotification.LargeIcon"/>
+        public string LargeIcon { get; set; }
+
+        /// <summary>
+        /// The picture to be displayed.
+        /// Can be resource name (like icon), file path or an URI supported by Android.
+        /// </summary>
+        public string Picture { get; set; }
+
+        /// <summary>
+        /// The content title to be displayed in the notification.
+        /// </summary>
+        public string ContentTitle { get; set; }
+
+        /// <summary>
+        /// The content description to set.
+        /// </summary>
+        public string ContentDescription { get; set; }
+
+        /// <summary>
+        /// The summary text to be shown.
+        /// </summary>
+        public string SummaryText { get; set; }
+
+        /// <summary>
+        /// Whether to show big picture in place of large icon when collapsed.
+        /// </summary>
+        public bool ShowWhenCollapsed { get; set; }
+    }
+
+    /// <summary>
     /// The AndroidNotification is used schedule a local notification, which includes the content of the notification.
     /// </summary>
     public struct AndroidNotification
@@ -69,6 +108,7 @@ namespace Unity.Notifications.Android
         /// Notification small icon.
         /// It will be used to represent the notification in the status bar and content view (unless overridden there by a large icon)
         /// The icon PNG file has to be placed in the `/Assets/Plugins/Android/res/drawable` folder and it's name has to be specified without the extension.
+        /// Alternatively it can also be URI supported by the OS.
         /// </summary>
         public string SmallIcon { get; set; }
 
@@ -91,6 +131,7 @@ namespace Unity.Notifications.Android
         /// Notification large icon.
         /// Add a large icon to the notification content view. This image will be shown on the left of the notification view in place of the small icon (which will be placed in a small badge atop the large icon).
         /// The icon PNG file has to be placed in the `/Assets/Plugins/Android/res/drawable folder` and it's name has to be specified without the extension.
+        /// Alternatively it can be a file path or system supported URI.
         /// </summary>
         public string LargeIcon { get; set; }
 
@@ -186,12 +227,30 @@ namespace Unity.Notifications.Android
             set => m_SilentInForeground = !value;
         }
 
+        /// <summary>
+        /// The necessary properties for big picture style notification.
+        /// For convenience, assigning this property will also set the Style property.
+        /// </summary>
+        public BigPictureStyle? BigPicture
+        {
+            get { return m_BigPictureStyle; }
+            set
+            {
+                m_BigPictureStyle = value;
+                if (m_BigPictureStyle.HasValue && Style == NotificationStyle.None)
+                    Style = NotificationStyle.BigPictureStyle;
+                else if (!m_BigPictureStyle.HasValue && Style == NotificationStyle.BigPictureStyle)
+                    Style = NotificationStyle.None;
+            }
+        }
+
         internal bool ShowCustomTimestamp { get; set; }
 
         private Color m_Color;
         private TimeSpan m_RepeatInterval;
         private DateTime m_CustomTimestamp;
         private bool m_SilentInForeground;
+        private BigPictureStyle? m_BigPictureStyle;
 
         /// <summary>
         /// Create a notification struct with all optional fields set to default values.
@@ -218,6 +277,7 @@ namespace Unity.Notifications.Android
             GroupAlertBehaviour = GroupAlertBehaviours.GroupAlertAll;
             ShowTimestamp = false;
             ShowCustomTimestamp = false;
+            m_BigPictureStyle = null;
 
             m_RepeatInterval = (-1L).ToTimeSpan();
             m_Color = new Color(0, 0, 0, 0);