+++ com.unity.mobile.notifications/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this package will be documented in this file.
 ## [Unreleased]
 
 ### Changes & Improvements:
+- [Android] - Added support for notification channel groups.
+- [Android] - Added support for BigPicture style.
+- [Android] - Icons now support local files and URIs.
 - [iOS] - Added support for notification action icons (requires iOS 15).
 
 ## [2.1.1] - 2023-01-04